move Core/IdeaColl/Poll to Core/Poll
