+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/IdeaCollection/Module.ts
@@ -1,15 +1,12 @@
-import * as AdhIdeaCollectionPollModule from "./Poll/Module";
 import * as AdhIdeaCollectionProcessModule from "./Process/Module";
 
 export var moduleName = "adhIdeaCollection";
 
 export var register = (angular) => {
-    AdhIdeaCollectionPollModule.register(angular);
     AdhIdeaCollectionProcessModule.register(angular);
 
     angular
         .module(moduleName, [
-            AdhIdeaCollectionPollModule.moduleName,
             AdhIdeaCollectionProcessModule.moduleName
         ]);
 };
