+++ src/adhocracy_frontend/adhocracy_frontend/static/js/Packages/Core/Poll/Detail.html
